+++ Server/src/main/java/com/tau/commstudy/controllers/UniversityController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.tau.commstudy.entities.University;
 import com.tau.commstudy.services.FacultyService;
 import com.tau.commstudy.services.UniversityService;
+import com.tau.commstudy.services.UserService;
 
 @RestController
 @RequestMapping("/university")
@@ -21,8 +22,12 @@ public class UniversityController {
     @Autowired
     private FacultyService facultyService;
 
+    @Autowired
+    private UserService userService;
+
     @RequestMapping(method = RequestMethod.POST, value = "/add")
-    public University add(@RequestBody University university) throws Exception {
+    public University add(@RequestBody University university, String userTokenId) throws Exception {
+	userService.assertAdminUser(userTokenId);
 	return universityService.add(university);
     }
 
@@ -32,12 +37,14 @@ public class UniversityController {
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/{id}")
-    public boolean delete(@PathVariable Long id) throws Exception {
+    public boolean delete(@PathVariable Long id, String userTokenId) throws Exception {
+	userService.assertAdminUser(userTokenId);
 	return universityService.delete(id);
     }
 
     @RequestMapping(method = RequestMethod.GET, value = "/addFaculty")
-    public boolean addFaculty(Long id, Long facultyId) throws Exception {
+    public boolean addFaculty(Long id, Long facultyId, String userTokenId) throws Exception {
+	userService.assertAdminUser(userTokenId);
 	return facultyService.addUniversity(facultyId, id);
     }
